Day6, part 2

+++ src/main/java/org/lecture/day6/Race.java
@@ -1,4 +1,4 @@
 package org.lecture.day6;
 
-public record Race(int time, int distance) {
+public record Race(long time, long distance) {
 }
